+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All user visible changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/), as described
 for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/text/1105-api-evolution.md)
 
+### Added
+
+* Added helper types for inner join and left outer join
+
 ## [0.15.1] - 2017-07-24
 
 * No changes to public API